+++ lib/puppet/functions/extended_stdlib/sqrt.rb
@@ -6,13 +6,14 @@ Puppet::Functions.create_function(:'extended_stdlib::sqrt') do
     param 'Numeric', :x
     return_type 'Float'
   end
-  # @param x The number of which to determine the square root.
+  # @param x The non-negative number of which to determine the square root.
   # @return Float The non-negative square root of the number.
   # @example Returns square root.
   #   sqrt(4) => 2.0
   #   sqrt(2) => 1.4142135623731
   #   sqrt(9) => 3.0
   def sqrt(num)
+    raise Puppet::Error, _('Input must be a non-negative number') if num.negative?
     Math.sqrt(num)
   end
 end