+++ spec/functions/sqrt_spec.rb
@@ -6,4 +6,5 @@ describe 'extended_stdlib::sqrt' do
   it { is_expected.to run.with_params(2).and_return(1.4142135623730951) }
   it { is_expected.to run.with_params(4).and_return(2.0) }
   it { is_expected.to run.with_params(9).and_return(3.0) }
+  it { is_expected.to run.with_params(-1).and_raise_error(Puppet::Error, _('Input must be a non-negative number')) }
 end